+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Business/Maturity/MaturityBusiness.cs
@@ -339,11 +339,24 @@ namespace CSETWebCore.Business.Maturity
 
 
         /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="assessmentId"></param>
+        /// <param name="isAcetInstallation"></param>
+        /// <param name="fill"></param>
+        /// <returns></returns>
+        public MaturityResponse GetMaturityQuestions(int assessmentId, bool isAcetInstallation, bool fill)
+        {
+            return GetMaturityQuestions(assessmentId, "*", isAcetInstallation, fill);
+        }
+
+
+        /// <summary>
         /// Assembles a response consisting of maturity settings for the assessment
         /// as well as the question set in its hierarchy of domains, practices, etc.
         /// </summary>
         /// <param name="assessmentId"></param>
-        public MaturityResponse GetMaturityQuestions(int assessmentId, bool isAcetInstallation, bool fill)
+        public MaturityResponse GetMaturityQuestions(int assessmentId, string domainAbbrev, bool isAcetInstallation, bool fill)
         {
             var response = new MaturityResponse();
