+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Helpers/ReportWidgets/ScoreBarChart.cs
@@ -63,7 +63,8 @@ namespace CSETWebCore.Helpers.ReportWidgets
                 xRect.SetAttributeValue("width", (barWidthInclGap - d.Gap).ToString());
                 xRect.SetAttributeValue("x", x.ToString());
                 xRect.SetAttributeValue("y", barTop.ToString());
-                xRect.SetAttributeValue("fill", d.BarColors[i]);
+                var fillColor = WidgetResources.ColorMap[d.BarColors[i]];
+                xRect.SetAttributeValue("fill", fillColor);
 
 
                 var xBarLabel = new XElement("text");
