+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Helpers/ReportWidgets/ScoreStackedBarChart.cs
@@ -46,7 +46,8 @@ namespace CSETWebCore.Helpers.ReportWidgets
                     xRect.SetAttributeValue("width", segmentWidth.ToString());
                     xRect.SetAttributeValue("x", x.ToString());
                     xRect.SetAttributeValue("y", "0");
-                    xRect.SetAttributeValue("fill", d.BarColors[i]);
+                    var fillColor = WidgetResources.ColorMap[d.BarColors[i]];
+                    xRect.SetAttributeValue("fill", fillColor);
 
 
                     // labels on the segments
@@ -58,7 +59,7 @@ namespace CSETWebCore.Helpers.ReportWidgets
                     xBarLabel.SetAttributeValue("text-anchor", "middle");
                     xBarLabel.SetAttributeValue("dominant-baseline", "middle");
                     xBarLabel.SetAttributeValue("class", "text-normal");
-                    var textColor = TextColorForBackground(d.BarColors[i]);
+                    var textColor = WidgetResources.GetTextColor(d.BarColors[i]);
                     xBarLabel.SetAttributeValue("fill", textColor);
 
                     x += segmentWidth;
@@ -74,38 +75,6 @@ namespace CSETWebCore.Helpers.ReportWidgets
         public override string ToString()
         {
             return xDoc.ToString();
-        }
-
-
-        /// <summary>
-        /// Calculates the luminance of the background color and returns
-        /// the contrasting text color.
-        /// </summary>
-        /// <param name="bgColor"></param>
-        /// <returns></returns>
-        private string TextColorForBackground(string bgColor)
-        {
-            // To change the cutoff, tweak this
-            var threshold = 0.179;
-
-            var color = bgColor.StartsWith('#') ? bgColor.Substring(1) : bgColor;
-            var r = Convert.ToInt32(color.Substring(0, 2), 16); // hexToR
-            var g = Convert.ToInt32(color.Substring(2, 2), 16); // hexToG
-            var b = Convert.ToInt32(color.Substring(4, 2), 16); // hexToB
-            var uicolors = new List<float>() { r / 255f, g / 255f, b / 255f };
-
-            var c = new List<float>();
-            foreach (float col in uicolors)
-            {
-                if (col <= 0.03928)
-                {
-                    c.Add(col / 12.92f);
-                }
-                c.Add((float)Math.Pow((double)((col + 0.055) / 1.055d), 2.4d));
-            };
-
-            var luminance = (0.2126 * c[0]) + (0.7152 * c[1]) + (0.0722 * c[2]);
-            return (luminance > threshold) ? "#000000" : "#FFFFFF";
-        }
+        }   
     }
 }
