+++ CSETWebNg/src/app/assessment/results/crr/crr-results-page/crr-results-page.component.html
@@ -2,7 +2,7 @@
     <div class="white-panel d-flex flex-column justify-content-start flex-11a">
         <h1 class="pageHeader">{{domain?.title}}</h1>
 
-        <app-crr-heatmap [domain]="domainId"></app-crr-heatmap>
+        <app-crr-heatmap [domainAbbrev]="domainAbbrev"></app-crr-heatmap>
 
         <div *ngIf="!loaded">
             <div class="spinner-container" style="margin: 2em auto">
@@ -10,7 +10,7 @@
             </div>
         </div>
 
-        <app-crr-results-detail *ngIf="loaded" [domain]="findDomain(domainId)"></app-crr-results-detail>
+        <app-crr-results-detail *ngIf="loaded" [domain]="domain"></app-crr-results-detail>
 
         <app-nav-back-next [page]="pageName"></app-nav-back-next>
     </div>
